type: add c_dvar_type_new_from_string() helper
Add new inline helper that creates a type from a full complete type
string. It simply falls back to the signature parser, but makes sure
the type-string does not exceed the first type.

This is only meant as helper for tests. You better fall back to the
signature parser for anything serious.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>