reader: implement skipping
Unfortunately, the D-Bus serialization does not support random access.
Hence, there is no way to easily skip parts of a variant. Instead, it
has to be parsed linearly.

As API user it is often preferably to skip unneeded parts, though.
Especially, skipping unwanted 'v' parts is very convenient. Hence, we
implement c_dvar_skip() for that. It takes the same format-string as the
reader, but additionally supports 'v'. It then skips over the variant,
regardless what type is stored in it.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>