var: make *_end_*() calls rewind variant
Instead of leaving the object untouched after *_end_*(), rewind the
variant for re-use. Regardless whether the caller needs it, this at
least has some value, rather than just placing it in an unused and
invalid state.

While at it, assert() on correct state to not trigger invalid behavior.
Writing to a ro variant, or reading from a writable variant should not
be left unassert'ed. It is invalid behavior and should better be caught
and fixed.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>