utf8: optimize the ASCII case
Most strings we parse are going to be ASCII, so let's try to make
that as fast as possible.

Using some bit-fiddling we can determine whether a given word contains
any NUL or non-ASCII bytes. Only when we find a word that does, do we
parse it byte-by-byte to find the first non-NULL, non-ASCII byte of
the substring.

This sped up my ASCII-only micro-benchmark by >30%.

This is inspired by Rust's std::str.

Signed-off-by: Tom Gundersen <teg@jklm.no>