test: don't evaluate higher end type macros
These are way to expensive to evaluate on lower-end machines. We
provide these for pre-calculation and testing of types. They're not
recommended to be used directly in your code.

Once we figure out a better way to utilize the preprocessor to compute
these, we might reconsider our decision. But it is much more likely
that we switch to a different language with better compile-time
computation features.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>