reader: prevent skipping '*' past array end
We currently allow using 'c_dvar_skip("*")` even if the iterator points
to the end of an array. This would then exit the array and thus trigger
an assertion on `depth > 0`. Fix this by catching this early and making
sure `*` can only be used to skip full types, never partial types.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>