c-dvar: fix out-of-bound reads on short buffers
Fix the c-dvar-reader byte fetchers to only access the data if the
return-code of the generic reader was 0. They used to do this also if it
was greater than zero (which also indicates errors and is wrong).

This was introduced way back in the initial implementation in:

    commit e4f1e356bdce04d3a34bb9ccb455945f7dabaa50
    Author: David Herrmann <dh.herrmann@gmail.com>
    Date:   Sat Apr 1 19:07:02 2017 +0200

        c-dvar: use fine-grained error codes

This commit changed the generic reader to return positive error-codes as
well as negative ones, but did not adjust the intermediate
read-helpers.

This bug allows accessing up to 8-bytes beyond the data-buffer and
reading it into a temporary variable. The data is not exposed, but the
bug can possibly trigger segmentation faults.

Reported-by: Evgeny Vereshchagin <evvers@ya.ru>
Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>