c-dvar: properly handle nested type-depth exhaustion
If types are nested via 'v', their maximum depth might exceed the
dbus-defined limits. We already check that this does not happen, but we
do not allow callers to properly handle this error. Make sure to return
it as C_DVAR_E_DEPTH_OVERFLOW just like the type-parser does.

This somewhat breaks the API, since callers must be aware of the code.
But since this was -ENOTRECOVERABLE before, callers need to adjust,
anyway.

Reported-by: Evgeny Vereshchagin <evvers@ya.ru>
Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>