reader: use c_load_*() to document aliasing
To read data from a buffer, use the new aliasing- and alignment-safe
helpers of c-stdaux. So far, we relied on the buffer to be owned by the
c-dvar implementation, and thus only ever be accessed by CDVar readers.
This is certainly not required, nor documented. By using `c_load_*()`,
aliasing is now allowed.

Aliasing on the buffer is unlikely to cause any actual miscompilations,
given that the buffer has a well-defined encoding that is unlikely to be
typed differently than intended. However, lets avoid the subtle API
constraints and just be aliasing-safe.

Reported-by: Kostadin Shishmanov <kocelfc@tutanota.com>
Signed-off-by: David Rheinsberg <david@readahead.eu>